requirements.txt
anyio>=4.11.0
asgiref>=3.10.0
bcrypt>=5.0.0
click>=8.3.0
colorama>=0.4.6
Django>=5.2.8
djangorestframework>=3.16.1
django-cors-headers>=4.9.0
dj-database-url>=3.0.1
django-filter>=25.2
dnspython>=2.8.0
gunicorn>=23.0.0
h11>=0.16.0
httptools>=0.7.1
idna>=3.11
Markdown>=3.10
mongoengine>=0.29.1
packaging>=25.0
pip>=25.3
pyjwt>=2.10.1
pymongo>=4.15.3
python-decouple>=3.8
python-dotenv>=1.2.1
PyYAML>=6.0.3
sniffio>=1.3.1
sqlparse>=0.5.3
tzdata>=2025.2
uvicorn[standard]>=0.38.0
watchfiles>=1.1.1
websockets>=15.0.1
whitenoise>=6.11.0
